🧹 refactor: extract CookieManager and AnalyticsManager to core
Extracted embedded classes from `content/components/footer/footer.js` into dedicated core modules:
- `content/core/cookie-manager.js`
- `content/core/analytics-manager.js`

This improves maintainability, readability, and allows for reusability of these utilities across the codebase. `SiteFooter` now imports these managers from the core directory.
