+++ content/core/analytics-manager.js
@@ -1,0 +1,55 @@
+import { createLogger } from './logger.js';
+
+const log = createLogger('AnalyticsManager');
+
+/**
+ * Analytics Manager
+ */
+export class AnalyticsManager {
+  #loaded = false;
+
+  load() {
+    if (this.#loaded) return;
+
+    document
+      .querySelectorAll('script[data-consent="required"]')
+      .forEach((script) => {
+        const newScript = document.createElement('script');
+        for (const attr of script.attributes) {
+          const name = attr.name === 'data-src' ? 'src' : attr.name;
+          if (!['data-consent', 'type'].includes(attr.name)) {
+            newScript.setAttribute(name, attr.value);
+          }
+        }
+        if (script.innerHTML.trim()) newScript.innerHTML = script.innerHTML;
+        script.replaceWith(newScript);
+      });
+
+    this.#loaded = true;
+    log.info('Analytics loaded');
+  }
+
+  /**
+   * @param {boolean} granted
+   */
+  updateConsent(granted) {
+    const win = /** @type {import('/content/core/types.js').GlobalWindow} */ (
+      window
+    );
+
+    if (typeof win.gtag !== 'function') return;
+
+    const status = granted ? 'granted' : 'denied';
+
+    try {
+      win.gtag('consent', 'update', {
+        ad_storage: status,
+        analytics_storage: status,
+        ad_user_data: status,
+        ad_personalization: status,
+      });
+    } catch (e) {
+      log.error('Consent update failed', e);
+    }
+  }
+}
