+++ content/core/cookie-manager.js
@@ -1,0 +1,53 @@
+/**
+ * Cookie Management Utility
+ */
+export const CookieManager = Object.freeze({
+  /**
+   * @param {string} name
+   * @param {string} value
+   * @param {number} [days=365]
+   */
+  set(name, value, days = 365) {
+    const date = new Date();
+    date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000);
+    const expires = `; expires=${date.toUTCString()}`;
+    const secure = window.location.protocol === 'https:' ? '; Secure' : '';
+    document.cookie = `${name}=${value || ''}${expires}; path=/; SameSite=Lax${secure}`;
+  },
+
+  /**
+   * @param {string} name
+   * @returns {string|null}
+   */
+  get(name) {
+    const nameEQ = `${name}=`;
+    const cookies = document.cookie.split(';');
+    for (let cookie of cookies) {
+      cookie = cookie.trim();
+      if (cookie.startsWith(nameEQ)) {
+        return cookie.substring(nameEQ.length);
+      }
+    }
+    return null;
+  },
+
+  /**
+   * @param {string} name
+   */
+  delete(name) {
+    const domains = [
+      '',
+      window.location.hostname,
+      `.${window.location.hostname}`,
+    ];
+    domains.forEach((domain) => {
+      const domainPart = domain ? `; domain=${domain}` : '';
+      document.cookie = `${name}=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/${domainPart}`;
+    });
+  },
+
+  deleteAnalytics() {
+    const analyticsCookies = ['_ga', '_gid', '_gat', '_gat_gtag_G_S0587RQ4CN'];
+    analyticsCookies.forEach((name) => this.delete(name));
+  },
+});
